feat(console): Implement a somewhat async logger
This is more of an experiment: Combining the `log` crate with async
code isn't easy, because of it's non-async API. As logging means cal-
ling non-async code, any async code that logs something loses it's
ability to `.await` at this point. It's possible, however, to run
more async code behind the logging API, provided it runs in a sepa-
rate executor. In theory, such an executor could be polled regularly
when the main thread has nothing to do (e.g. delays, async i/o).

For a simple polling loop that forwards data from a pipe like in this
example, `async` is overkill though. One could simply expose a single-
ton poll function for this.

Also, so far no ready, non-alloc, safe-to-use executor could be found
for this case.

Signed-off-by: Nico Huber <nico.h@gmx.de>